chore: organize components and services
