+++ components/UserInputPanel.tsx
