+++ services/geminiService.ts
